CP-47935: Make `import_file.py` as a separate module
- Make `import_file.py` as a separate module for reusing it.
- Fix pylint and pytype issues
- No functionality change
